move LinkButton to common, use in geo
